build: Workaround AOSP DT modification to DTBs
AOSP ROM device trees for the Galaxy A50 omits the product partition in its DTBs. Since this is mostly unused (GSIs or otherwise) but the way we build requires the DTB to be in a specific directory, workaround this by copying a version for AOSP when we are building automated builds (CI).

We'll upstream changes to this ourselves.

Signed-off-by: John Vincent <git@tenseventyseven.cf>