arm64: dts: exynos9610: Tune frequency variant (freqvar) scaling
Initial attempts to better scale frequencies on the big and LITTLE clusters than stock behavior, especially that we have upstreamed and removed any sort of "AI"/random scaling. This is largely targeted to the big cluster which weirdly wants to stay at the lowest regardless of CPU load.

While we're at it, revert "Re-adjust FRT ratios" on AOSP DTS and sync changes to it.

Signed-off-by: John Vincent <git@tenseventyseven.cf>