[skip rel] arm64: dts: exynos9610: Use most efficient frequency on the big cluster as minimum
We might need to re-measure for the Exynos 9610 but measurements done for the Exynos 9611 has shown that 1144Mhz has shown to be the most efficient frequency on the big cluster[1][2], but this causes thermal issues.[3]

Set the minimum to be the next step down, 1040Mhz. Reflect the change on our Freqvar Tune boost and upscale tables.

[1] https://github.com/Dark-Matter7232/Cosmic-fresh/commit/53935a6b04612f9fe092aca9820f4b9e02e64fe6
[2] https://github.com/kdrag0n/freqbench/blob/master/results/exynos9611/main/results.csv
[3] https://github.com/Dark-Matter7232/Cosmic-fresh/commit/9485725dbf64c77f0d5d83672d61102605c034bc
[4] https://github.com/Panchajanya1999/msm-4.14/commit/b45fdffb230d67e175430a8a349cebaeaadee816

Co-authored-by: Dark-Matter7232 <kerneldeveloper7232@gmail.com>
Signed-off-by: John Vincent <git@tenseventyseven.cf>